// packages/cmd-hub/src/db/models-enum.ts
export enum DbModelsEnum {
    Managers = "managers",
    Accounts = "accounts",
    AccountModules = "account_modules",
    AccountSessions = "account_sessions",
    Files = "files",
    DefaultAssets = "default_assets",
    MsgHistory = "message_history",
    CmdAliases = "cmd_aliases",
    PendingDeletes = "pending_deletes",
    SystemConfigs = "system_configs",
    UserConfigs = "user_configs",
}